Init commit - refactor task to better support common elements and new paradigms
