+++ bcipy/task/README.md
@@ -6,7 +6,7 @@ These are the experimental tasks that can be implemented.
 ## Modes
 ---------
 Within `tasks/` there are folders for each of the supported modes, and within them, the supported experiment types. To add new modes, create a folder for it and place the tasks in files within it. Be sure to add it to the `start_task` file at the root to be able execute it! An entry must also be added to the task_registry TaskType
-enum. This updates the GUI (RSVPKeyboard.py) as well makes the task available to `start_task`.
+enum. This updates the GUI (BCInterface.py) as well makes the task available to `start_task`.
 
 Currently, these are the modes and experiment types implemented:
 
@@ -14,8 +14,6 @@ Currently, these are the modes and experiment types implemented:
 
 > Calibration
 > Copy Phrase
-> Icon to Icon
-> Icon to Word
 > Alert Tone
 > Inter-Inquiry Feedback Calibration
 
@@ -26,27 +24,20 @@ Currently, these are the modes and experiment types implemented:
 *Start Task* 
 
 Start Task takes in Display [object], parameters [dict], file save [str-path] and task type [dict]. Using the
-task type, start_task() will route to the correct mode (RSVP, SSVEP, MATRIX) and experiment type (Calibration, Copy Phrase, etc.)
+task type, start_task() will route to the correct paradigm (RSVP, SSVEP, MATRIX) and mode (Calibration, Copy Phrase, etc.)
 
 It is called in the following way:
 
+
 ```
-	from bci_tasks.start_task import start_task
+	from bcipy.task import start_task
 
     start_task(
        	display_window,
-        task_type,
+        data_acquisition_client,
         parameters,
         file_save)
 
 ```
 
-It will throw an error that the task isn't implemented if given a mode or experiment type that's unavailable to it. 
-
-
-## Using Images
----------------
-
-1. You make a folder in the `static/images` dir with your .pngs
-2. Point to that folder in the parameters.json file
-3. Set is_txt_stim to false
+It will throw an error that the task isn't implemented.
