+++ bcipy/task/__init__.py
@@ -1,0 +1,8 @@
+"""
+This import statement allows users to import submodules from Task
+"""
+from .main import Task
+
+__all__ = [
+    'Task',
+]
