+++ bcipy/task/control/criteria.py
